+++ reference/programming-languages/scala/modules/scala/functional/programming/book/package-info.java
@@ -1,0 +1,1 @@
+package scala.functional.programming.book;